Move tag definitions out of main syntax file
